requirements.txt
# basics
pytorch-lightning==1.9.5
torchmetrics==1.2.0 
einops==0.7.0
editdistance==0.6.0
pillow==9.2.0
opencv-python==4.8.1.78
# for pytorch-lightning CLI
jsonargparse[signatures]==4.25.0
# export
onnx==1.13.0
# dev-dependency
flake8==6.1.0
black==23.9.1
isort==5.12.0
ipykernel==6.25.2
notebook==7.0.4
# for test
typer==0.6.1
beautifulsoup4==4.11.1
lxml>=4.9.1
